Partially revert "Enable Libdl in PCRE"
This reverts most commit 3bd137c40b0b0eddea46e48347a82bfa3570838c.

Most changes to path.jl and file.jl are not reverted.
